feat: fork delta rebased onto upstream/main (virtual library support)
Replay of the drondeseries fork delta on top of Silo-Server/silo-server
upstream/main@9074aeab, so origin/main is 0 commits behind upstream.

Kept (fork features): virtual library support end to end (virtual://
sources, plugin stream resolution and relay, canonical-input recipe cards,
session virtual-source binding, candidate failover/unpin healing, virtual
purge + collection sync); tonemap source validation for remote/virtual
inputs; Apple AVPlayer MKV container quirk; scanner bitrate-completeness
repair; ProbeVirtualSource; jellycompat virtual download/binding flows and
shouldUseCompatNodePool gating; device profile persistence; web detail hover
polish, Delete Item menu entry, and synced-collection theatrical gate.

Adopted from upstream (through 9074aeab): playback startup hardening;
audio-v2 stereo-downmix remux; tokenless V3 claims, node-recipe/proxy-grant
stores and authorized-origin egress; DV7 client-managed dynamic range;
HDR-to-SDR tone mapping evolution; Safari HLS remux; scanner DV reprobe;
copy-safety verdict persistence; stream telemetry; image_size artwork ladder
+ backfill; season/episode browse-latency work; watched indicators and
poster quick actions; migrations. Feature changelog removed per upstream.

Deliberately not adopted: optimistic copy-safety start-path race +
plan_invalidated_v1 client plumbing; proxy /stream/v3 grant routes remain
unmounted.

AI-assisted merge resolution; see PR body for evidence.

+++ internal/playback/directplay_revision_linux.go
@@ -15,7 +15,7 @@ func directPlayFilesystemRevision(_ *os.File, info os.FileInfo) (string, bool) {
 	}
 	return fmt.Sprintf(
 		"linux:%x:%x:%x:%x:%x:%x",
-		uint64(stat.Dev),
+		stat.Dev,
 		stat.Ino,
 		stat.Ctim.Sec,
 		stat.Ctim.Nsec,
